+++ kotti_settings/events.py
@@ -1,0 +1,14 @@
+
+
+class SettingsEvent(object):
+    """ """
+    def __init__(self, module=None):
+        self.module = module
+
+
+class SettingsBeforeSave(SettingsEvent):
+    pass
+
+
+class SettingsAfterSave(SettingsEvent):
+    pass